# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Nunu Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/v1/team/addTeamMember": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "添加团队成员",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddTeamMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/applyJoinTeam": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "申请加入团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ApplyJoinTeamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ApplyJoinTeamResp"
                        }
                    }
                }
            }
        },
        "/v1/team/createTeam": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "新建团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateTeamResp"
                        }
                    }
                }
            }
        },
        "/v1/team/deleteTeam": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "删除团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/deleteTeamMember": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "删除团队成员",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTeamMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/getTeamApplyList": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "获取团队申请列表（个人/管理）",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamApplyListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamApplyListResp"
                        }
                    }
                }
            }
        },
        "/v1/team/getTeamInfo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "获取团队详细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamInfoResp"
                        }
                    }
                }
            }
        },
        "/v1/team/getTeamList": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "获取团队列表",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamListResp"
                        }
                    }
                }
            }
        },
        "/v1/team/getTeamMemberList": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "获取团队成员列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetTeamMemberListResp"
                        }
                    }
                }
            }
        },
        "/v1/team/getUserTeamList": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "获取个人团队列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Index",
                        "name": "pageIndex",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetUserTeamListResp"
                        }
                    }
                }
            }
        },
        "/v1/team/handleTeamApply": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "处理团队申请",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.HandleTeamApplyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/quitTeam": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "退出团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.QuitTeamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/updateTeam": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "更新团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/updateTeamMemberRole": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队模块"
                ],
                "summary": "更新团队成员角色",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateTeamMemberRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "v1.AddTeamMemberReq": {
            "type": "object",
            "properties": {
                "member_id": {
                    "description": "成员ID",
                    "type": "string"
                },
                "role_type": {
                    "description": "角色类型",
                    "type": "string"
                },
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
        "v1.ApplyJoinTeamReq": {
            "type": "object",
            "properties": {
                "reason": {
                    "description": "申请理由",
                    "type": "string"
                },
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
        "v1.ApplyJoinTeamResp": {
            "type": "object",
            "properties": {
                "apply_id": {
                    "description": "申请ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.CreateTeamRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "team_name": {
                    "description": "团队名",
                    "type": "string"
                }
            }
        },
        "v1.CreateTeamResp": {
            "type": "object",
            "properties": {
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.DeleteTeamMemberReq": {
            "type": "object",
            "properties": {
                "member_id": {
                    "description": "成员ID",
                    "type": "string"
                },
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
        "v1.DeleteTeamRequest": {
            "type": "object",
            "properties": {
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.GetTeamApplyListReq": {
            "type": "object",
            "properties": {
                "is_personal": {
                    "description": "是否是个人申请",
                    "type": "boolean"
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "status": {
                    "description": "申请状态",
                    "type": "integer"
                },
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
        "v1.GetTeamApplyListResp": {
            "type": "object",
            "properties": {
                "apply_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.TeamApplyData"
                    }
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "v1.GetTeamInfoResp": {
            "type": "object",
            "properties": {
                "member": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.MemberData"
                    }
                },
                "team": {
                    "$ref": "#/definitions/v1.TeamData"
                }
            }
        },
        "v1.GetTeamListReq": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "创建者ID",
                    "type": "string"
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "team_name": {
                    "description": "团队名",
                    "type": "string"
                }
            }
        },
        "v1.GetTeamListResp": {
            "type": "object",
            "properties": {
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "team_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.TeamData"
                    }
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "v1.GetTeamMemberListResp": {
            "type": "object",
            "properties": {
                "member_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.MemberData"
                    }
                }
            }
        },
...
        "v1.GetUserTeamListResp": {
            "type": "object",
            "properties": {
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "team_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.TeamData"
                    }
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "v1.HandleTeamApplyReq": {
            "type": "object",
            "properties": {
                "apply_id": {
                    "description": "申请ID",
                    "type": "integer"
                },
                "status": {
                    "description": "申请状态",
                    "type": "integer"
                },
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.QuitTeamReq": {
            "type": "object",
            "properties": {
                "team_id": {
                    "description": "团队ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}